refactor(phase-5): fold config.py into env + split webui/app.py
Phase 5 increment covering the first two PRs of the tracking issue:

1. Fold root config.py (DEBUG flag) into the MICHEL_DEBUG env var.
   - src/core/logging.py now reads DEBUG from os.environ instead of
     importing a top-level config module.
   - main.py re-exports DEBUG through src.core.logging.
   - Root config.py is deleted.

2. Split src/webui/app.py (1016 LOC) into focused route modules.
   - src/webui/context.py: WebUIContext dataclass with shared helpers
     (config I/O, session/auth enforcement, extension discovery).
   - src/webui/routes/auth.py: /auth/* + /api/me.
   - src/webui/routes/config.py: /api/config, /api/modules,
     /api/schemas/*, /api/global-config/*, /api/cleanup-config.
   - src/webui/routes/servers.py: /api/servers and per-server channels,
     members, module update/toggle, EmbedManager publish.
   - src/webui/routes/extensions.py: /api/extensions + reload.
   - src/webui/routes/bot.py: /api/bot-info.
   - src/webui/routes/frontend.py: /, /health, and SPA catch-all.
   - src/webui/sse/logs.py: /api/logs and the SSE log stream.
   - src/webui/app.py is now ~72 lines: build context + mount routers.

Each router exposes a create_router(ctx) factory that binds routes to
the shared WebUIContext, replacing the closure-heavy helpers inside the
old monolithic factory.

Verified: ruff clean; create_app() boots with the same 28 endpoints as
before (plus the default /openapi.json).

Refs #15 (Phase 5).
