--- config.py
@@ -1,3 +1,0 @@
-"Enable DEBUG messages for logging"
-
-DEBUG = False
