+++ src/webui/routes/__init__.py
@@ -1,0 +1,5 @@
+"""Web UI route modules, split by domain.
+
+Each module exposes a ``create_router(ctx)`` factory that binds routes to the
+shared :class:`~src.webui.context.WebUIContext`.
+"""
