+++ src/webui/sse/__init__.py
@@ -1,0 +1,1 @@
+"""Server-Sent Events endpoints for the Web UI (log streaming, etc.)."""
